0.3.3 tested final version

+++ src/version_1/Loader.java
@@ -34,6 +34,7 @@ public class Loader
                 {
                     data = null;
                 }
+
                 if (Objects.equals(type[i], "Skip"))
                 {
                     continue;
@@ -176,7 +177,7 @@ public class Loader
                 database.close(stmt);
                 long end = System.currentTimeMillis();//结束时间
                 System.out.println(cnt + " records successfully loaded");
-                System.out.println("TIME : " + (long) (end - start) / 1000 + "s");
+                System.out.println("TIME : " + (end - start) / 1000 + "s");
                 System.out.println("Loading speed : " + (long) cnt / ((end - start) / 1000) + " records/s");
             }
             catch (Exception e)
